refactor(cli): consolidate config-filename + rule-name-set helpers
Closes #68. Three near-duplicates folded into shared layers:

- `config.Filenames` now owns the `["krit.yml", ".krit.yml"]` list;
  `internal/lsp`, `internal/cli/scan/early_exits`, and
  `internal/projectroot/root` reference it instead of inlining the slice.
- `clishared.FindConfigInDir(dir)` replaces the daemon's `findRepoConfig`
  and the bulk of `detectConfigForScanArgs`.
- `clishared.ParseRuleNameSetCSV` replaces the daemon's
  `parseRuleNameSetCSV` and the CLI's unexported `parseRuleNameSet`.
